Clear auth state on auth failure (#4341)
* Clear all previous auth state on 401

* Update docs on medplumClient

* Prevent calling sessionStorage in Node env

+++ packages/core/src/client.ts
@@ -906,7 +906,9 @@ export class MedplumClient extends EventTarget {
    */
   clear(): void {
     this.storage.clear();
-    sessionStorage.clear();
+    if (typeof window !== 'undefined') {
+      sessionStorage.clear();
+    }
     this.clearActiveLogin();
   }
 
@@ -3310,7 +3312,7 @@ export class MedplumClient extends EventTarget {
     if (this.refresh()) {
       return this.request(method, url, options);
     }
-    this.clearActiveLogin();
+    this.clear();
     if (this.onUnauthenticated) {
       this.onUnauthenticated();
     }
